bugs/serializers.py
from rest_framework import serializers
from .models import Bug

class BugSerializer(serializers.ModelSerializer):
    class Meta:
        model = Bug
        fields = ('bug_title', 'pub_date', 'bug_text')
