pyproject.toml
[tool.poetry]
name = "BugTracker"
version = "0.1.0"
description = ""
authors = ["yankihue <yankisirtas@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
